requirements.txt
# CryptoKit (CK) Requirements
# Core cryptographic libraries
cryptography>=41.0.0
pycryptodome>=3.19.0

# CLI and UI libraries
click>=8.1.0
rich>=13.6.0
colorama>=0.4.6

# File and data processing
Pillow>=10.0.0
PyYAML>=6.0.1
python-magic>=0.4.27

# System utilities
psutil>=5.9.0
pathlib2>=2.3.7

# Hashing and analysis
hashlib-compat>=1.0.1
python-blake2>=0.1.1
